requirements.txt
Django==3.0.8
requests==2.24.0
pilkit==2.0
Pillow==7.2.0
django-imagekit==4.0.2